+++ user-service/src/main/kotlin/org/ult1mma/userservice/service/UserService.kt
@@ -30,6 +30,13 @@ class UserService(private val userRepository: UserRepository) {
 
     fun create(user: User): User {
         logger.info("Создание пользователя: {}", user)
+        if (userRepository.existsByEmail(user.email)) {
+            logger.warn("Пользоваттель с таким email уже существует {}", user.email)
+            return user
+        }
+        if (userRepository.existsByName(user.name)) {
+            logger.warn("Пользователь с таким name(именем) уже существует {}", user.name)
+        }
         val saved = userRepository.save(user)
         logger.info("Пользователь успешно создан: id={}", saved.id)
         return saved
@@ -37,6 +44,10 @@ class UserService(private val userRepository: UserRepository) {
 
     fun delete(id: Long) {
         logger.info("Удаление пользователя с id={}", id)
+        if (userRepository.existsById(id).not()) {
+            logger.warn("Пользователь с таким id={} не найден", id)
+            return
+        }
         userRepository.deleteById(id)
         logger.info("Пользователь с id={} удалён", id)
     }
